Pushed the new Odin Unity SDK 2.0 preview version
- Contains the Core SDK Version 2.0.0-beta-crypto2
- Now supports WebGL Platform
- Contains specific WebGL Sample Scene

+++ Runtime/Events/UnityDeleteMediaObject.cs
@@ -1,10 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnityEngine.Events;
-using OdinNative.Odin.Room;
 
 namespace OdinNative.Unity.Events
 {
@@ -12,7 +7,7 @@ namespace OdinNative.Unity.Events
     /// Unity Inspector event wrapper <see href="https://docs.unity3d.com/ScriptReference/Events.UnityEvent.html">(UnityEvent)</see>
     /// </summary>
     [Serializable]
-    public class UnityDeleteMediaObject : UnityEvent<long>
+    public class UnityDeleteMediaObject : UnityEvent<ushort>
     {
     }
 }
